Memoization for NDimen and Switch to Naive1D
Profiling revealed that 40+% of CPU time was spent on generating power
sets in NDimen. This was slightly improved through memoization;
switching from NDimen to Naive1D improved run-time by 2+ orders of
magnitude. The experiment should be modified to take in a real dataset
(not random data), but this would not change the runtime.